fix(SLI): ki_tile 保 fp16 进 tl.dot，省 32KB UB 修复 9382 溢出
G=32 之后单核 UB 仍超 16KB，主因是 ki_tile [128,128] fp32 = 64KB 经
multi-buffer 翻倍占 128KB。把 ki_tile 保持 fp16，ds_idx cast 到 fp16
后做 fp16 x fp16 -> fp32 dot，UB 立省 32KB。

副作用：cube 半精度通路吞吐更高，单 chunk 耗时 1947us → 972us，比
之前 G=64 fp32 dot 的 995us 还略快，相对原始 6452us 共 6.6x 提升。
本地 910B3 + CANN 8.5.1 smoke pass。

+++ sparse_lightning_indexer_grad_kl_loss_triton.py
@@ -389,14 +389,17 @@ def _query_index_weight_grad_kernel(
         ds_idx = di_tile[None, :] * w_g[:, None] * relu_mask
 
         # ki_tile: [BLOCK_K, BLOCK_D], shared across all g
+        # Keep ki_tile in fp16 to halve UB pressure on tight backends
+        # (e.g. CANN 9.0 + Ascend910_9382 multi-buffer pass). tl.dot
+        # accepts fp16 x fp16 -> fp32 accumulator.
         ki_tile = tl.load(
             key_index_gathered_ptr + ki_g_base
             + k_offs[:, None] * D_idx + d_offs[None, :],
             mask=k_mask[:, None] & d_valid[None, :],
-            other=0.0).to(tl.float32)
+            other=0.0)
 
         # dqi_acc += [BLOCK_G, BLOCK_K] @ [BLOCK_K, BLOCK_D]
-        dqi_acc += tl.dot(ds_idx, ki_tile)
+        dqi_acc += tl.dot(ds_idx.to(ki_tile.dtype), ki_tile)
 
     # Store dQueryIndex tile [BLOCK_G, BLOCK_D]
     dqi_offs = (qi_base
